Rename package from identitypool to authentication
