+++ mqtt-impl/src/test/java/io/streamnative/pulsar/handlers/mqtt/authentication/AuthenticationProviderMTlsTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.mqtt.identitypool;
+package io.streamnative.pulsar.handlers.mqtt.authentication;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.netty.channel.local.LocalAddress;
